requirements.txt
certifi==2018.10.15
chardet==3.0.4
dataclasses==0.6
Django==2.1.2
django-filter==2.0.0
django-rest-auth==0.9.3
djangorestframework==3.9.0
djongo==1.2.30
gcloud==0.17.0
googleapis-common-protos==1.5.3
httplib2==0.11.3
idna==2.7
jws==0.1.3
Markdown==3.0.1
oauth2client==3.0.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycryptodome==3.4.3
pymongo==3.7.2
python-jwt==2.0.1
pytz==2018.5
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.11.0
sqlparse==0.2.4
urllib3==1.24
